+++ docs/about/changelog.md
@@ -19,6 +19,7 @@ toc_depth: 2
 - Fix `json.set` nx arg behavior #372
 - Fix issue with `hset` not resetting expiration over field
 - Fix inconsistent 'pending' field in return from XINFO GROUPS #374
+- Fix `TDIGEST.QUANTILE` to match redis behavior on a TDigest with one value #377
 
 ## v2.28.1 - 2025-04-02